Frame Extractor
Moves Frame Extractor code into it's own frame_extractor.py file.
